+++ src/lib/parseConfig.ts
@@ -3,33 +3,22 @@ import path from "path";
 import { Color } from "./colorSvg.js";
 
 export type Config = {
-  dir: string;
-  out: string;
-  fps?: number;
+  use: "puppeteer" | "default";
   colors?: Color[];
+  dir: string;
+  fps?: number;
+  out: string;
 };
 
 export type Configs = {
   [key: string]: Config;
 };
 
-export type ParsedConfig = {
-  use: "puppeteer" | "default";
-  configs: Configs;
-};
-
-export const parseConfig = (p: string): ParsedConfig => {
+export const parseConfig = (p: string): Configs => {
   const file = path.resolve(p);
   const configs = JSON.parse(
     fs.readFileSync(file, { encoding: "utf-8", flag: "r" }),
   );
 
-  const use = configs["use"];
-  delete configs["use"];
-
-  if (use.lowerCase() === "puppeteer") {
-    return { configs, use: "puppeteer" };
-  } else {
-    return { configs, use: "default" };
-  }
+  return configs;
 };